requirements.txt
numpy>=1.19.5
opencv-python-headless>=4.5.3
# tensorflow==2.15.0  # Commented out as we'll use tflite-runtime instead
tflite-runtime>=2.5.0
pillow>=8.0.0
pyserial>=3.5
# Additional dependencies for Raspberry Pi
RPi.GPIO>=0.7.0  # For GPIO access if needed